% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/careless-package.R
\docType{data}
\name{bfi}
\alias{bfi}
\title{25 Personality items representing 5 factors}
\format{A data frame with 2,800 rows and 28 varialbes:
\describe{
   \item{A1:O5}{Raw item responses on 25 items, 5 per personality scale}
   \item{gender}{Factor with 2 levels, male = 1 and female = 2}
   \item{education}{Number of years of high school education, range 1-5}
   \item{age}{Age of participant, range 3-86}
 }}
\source{
\url{http://cran.r-project.org/web/packages/psych/index.html}
}
\usage{
bfi
}
\description{
A dataset containing 25 personality self-report items taken from the International Personality
Item Pool (ipip.ori.org), and packaged with the \code{psych} library in R. Included here
for demonstrative purposes; see the vignette for more details.
}
\keyword{datasets}

