% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/careless-package.R
\docType{package}
\name{careless-package}
\alias{careless-package}
\title{Metrics for identifying careless responding.}
\description{
careless: Metrics for identifying careless responding.
}
\details{
The careless package provides a variety of metrics for identifying careless responding.
This these are drawn from the statistics used by Craig and Meade (2012).
}
\section{careless functions}{
 So far, this package implements the following functions:
`longString()`, which returns either the count or the value of the longest set of
repititions in a dataframe.
}
\author{
Matthew Sigal \email{matthewsigal@gmail.com}
}
\keyword{package}

