% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/careless.R
\docType{package}
\name{careless}
\alias{careless}
\alias{careless-package}
\title{careless: A package for identifying careless responding.}
\description{
The careless package provides a variety of metrics for identifying careless responding.
Inititally, these are drawn from the statistics used by Craig and Meade (2012).
}
\section{careless functions}{
 The careless functions ...
}

