% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careless_dataset2.R
\docType{data}
\name{careless_dataset2}
\alias{careless_dataset2}
\title{Simulated dataset with careless responses.}
\format{
A data frame with 1000 observations (rows) and 100 variables (columns).
}
\usage{
careless_dataset2
}
\description{
A simulated dataset mimicking insufficient effort responding. Contains three types of responses:
(a) Normal responses with answers mimicking a diligent respondent
(b) Some number of longstring careless responders,
(c) some number of generally careless responders.
Simulated are 10 subscales of 10 items each (= 100 variables).
}
\keyword{datasets}
