% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careless_testing_na.R
\docType{data}
\name{careless_testing_na}
\alias{careless_testing_na}
\title{Simulated dataset with insufficient effort responses.}
\format{
A data frame with 200 observations (rows) and 50 variables (columns).

A datset created using careless_dataset as a base with the following
modifications:
1) the 103 row is entirely NA
2)  the 144th row's first 10 values are NA
3) the 138th row's last 20 values are NA
4) each value in the 2nd row had a 5% chance of being NA
5) Row 194 has 1 NA
}
\usage{
careless_testing_na
}
\description{
A simulated dataset mimicking insufficient effort responding. Contains three types of responses:
(a) Normal responses with answers centering around a trait/attitude value
(80 percent probability per simulated observation),
(b) Straightlining responses (10 percent probability per simulated observation),
(c) Random responses (10 percent probability per simulated observation).
Simulated are 10 subscales of 5 items each (= 50 variables).
}
\keyword{datasets}
