% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carelessDataset.R
\docType{data}
\name{carelessDataset}
\alias{carelessDataset}
\title{Simulated dataset with insufficient effort responses.}
\format{A data frame with 200 observations (rows) and 50 variables (columns).}
\usage{
carelessDataset
}
\description{
A simulated dataset mimicking insufficient effort responding. Contains three types of responses:
(a) Normal responses with answers centering around a trait/attitude value
(80 percent probability per simulated observation),
(b)Straightlining responses (10 percent probability per simulated observation),
(c) Random responses (10 percent probability per simulated observation).
Simulated are 10 subscales of 5 items each (= 50 variables).
}
\keyword{datasets}
