% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/psySyn.R
\name{countSyn}
\alias{countSyn}
\title{Count Psychological Synonyms}
\usage{
countSyn(x, cutoff = 0.6, cor.method = "pearson")
}
\arguments{
\item{x}{An \code{R} dataframe or matrix object.}

\item{cutoff}{A numeric scalar. The value at which correlations should be declared "synonyms".
Default is set to \emph{r} >= +.60.}

\item{cor.method}{Character string indicating which method is passed to \code{cor} to use when
calculating the correlation coefficients.}
}
\value{
A list detailing the number of synonyms found, and their values.
}
\description{
\code{countSyn} prints the number of Psychological Synonyms found at a particular cut-off,
as well as the variables they pertain to, and their correlation.
}
\seealso{
\code{\link{careless}}, \code{\link{longString}}, \code{\link{psyAnt}}, \code{\link{malDist}}
}

