% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longstring.R
\name{longstring}
\alias{longstring}
\title{Identifies the longest string of identical responses for each observation}
\usage{
longstring(data, na = F)
}
\arguments{
\item{data}{is a matrix of data}

\item{na}{is a boolean indicating whether na values are present in the data}
}
\description{
Takes a matrix of item responses and, beginning with the second column 
compares each column with the previous one to check for matching responses.
For each observation, the length of the maximum uninterupted chain of 
identical repsonses is returned.
}
\examples{
\dontrun{
d <- matrix(rep(1:10,10), ncol=10, nrow=10, byrow=T)
for(i in 1:10){
  d[i,which(1:10 <= i)] <- i
}
longstring(d)
}
Last Updated 14-June-2016
}
\author{
Richard Yentes \email(rdyentes@ncsu.edu)
}

