% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core-functions.R
\name{LSmax}
\alias{LSmax}
\title{Long string max.}
\usage{
LSmax(myRow)
}
\arguments{
\item{...}{Dataframe}

\item{na.rm}{A logical scalar. Should missing values (including NaN) be removed?}
}
\value{
A vector of type integer.
}
\description{
\code{LSmax} returns the sum of repeated values.
}
\details{
This function...
}
\examples{
\dontrun{
LSmax(sample(1:4, 30, replace = TRUE))

}
}
\seealso{
\code{\link{functionname}}, \code{\link[packagename]{functioname}}

Other careless.responding.functions: \code{\link{psy.syn}}
}

