% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/malDist.R
\name{malDist}
\alias{malDist}
\title{Mahalanobis Distance.}
\usage{
malDist(x)
}
\arguments{
\item{x}{An `R` dataframe or matrix object.}
}
\value{
A vector of type numeric.
}
\description{
\code{malDist} returns a vector indicating the Mahalanobis Distance for each participant in a
dataframe.
}
\details{
This function is designed for univariate tests, and assumes that all variables passed belong
to the same factor. Also, if data are passed as factors, it is converted to integer before
calculating the Mahalanobis Distance. If missing data is present, the function will return
NA for that observation.
}
\examples{
\dontrun{
set.seed(77)
dat <- matrix(sample(1:3, 200, replace = TRUE),20,10)
malDist(dat)
}
}
\seealso{
\code{\link{longString}}
}

