% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Evaluates each row of a matrix for outlier status.}
\usage{
outlier(data, confidence = 0.95, raw = FALSE)
}
\arguments{
\item{data}{is a matrix of data}

\item{confidence}{is the desired confidence level of the result 
(default =.95)}

\item{raw}{determines whether the result vector contains outlier flags or raw
Mahalanobis distance values (default = FALSE)}
}
\description{
Takes a matrix of item repsonses and computes Mahalanobis distance. Returns a
vector of binary outlier flags by default. If the raw options is specified 
the function returns a vector of raw Mahalanobis distance scores.
}
\examples{
\dontrun{
x <- outlier(data)
mahd <- outlier(data,raw=T)
z <- outlier(data, confidence=.999)
}
}
\author{
Richard Yentes \email{rdyentes@ncsu.edu}
}
