% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarityMeasures.R
\name{psy.syn}
\alias{psy.syn}
\title{Psychological Synonyms.}
\usage{
psy.syn(input_data, test_value)
}
\arguments{
\item{input_data}{Dataframe}

\item{test_value}{A logical scalar. Should missing values (including NaN) be removed?}
}
\value{
A vector of type integer.
}
\description{
\code{psy.syn} returns...
}
\details{
This function...
}
\seealso{
\code{\link{functionname}}, \code{\link[packagename]{functioname}}

Other careless.responding.functions: \code{\link{longString}}
}

