% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/psyAnt.R
\name{psyAnt}
\alias{psyAnt}
\title{Psychological Antonyms}
\usage{
psyAnt(x, cutoff = -0.6, cor.method = "pearson")
}
\arguments{
\item{x}{An \code{R} dataframe or matrix object.}

\item{cutoff}{A numeric scalar. The value at which correlations should be declared "antonyms".
Default is set to \emph{r} <= -.60.}

\item{cor.method}{Character string indicating which method is passed to \code{cor} to use when
calculating the correlation coefficients.}
}
\value{
A vector of type numeric.
}
\description{
\code{psyAnt} returns a numeric vector, pertaining to the individual's consistency index based
upon the correlation between items of a small magnitude.
}
\details{
This function is a consistency index based upon correlating all items within a survey, finding
the items with the smallest correlations, and then calculating the within item correlation
among those measures.
}
\seealso{
\code{\link{careless}}, \code{\link{longString}}, \code{\link{psySyn}}, \code{\link{malDist}}
}

