% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychsyn.R
\name{psychsyn}
\alias{psychsyn}
\title{Computes the psychometric synonym score}
\usage{
psychsyn(data, critVal = 0.6, anto = FALSE)
}
\arguments{
\item{data}{is a matrix of item responses}

\item{critVal}{is the minimum magnitude of the correlation between two items
in order for them to be considered psychometric synonyms. Defaults to .60}

\item{anto}{determines whether psychometric anonyms are returned instead of 
psychometric synonyms. Defaults to FALSE}
}
\description{
Takes a matrix of item responses and identifies item pairs that are highly 
correlated within the overall dataset. Each respondents psychometric synonym
score is then comptued as the within-person correlation between the
identified item-pairs.
}
\examples{
\dontrun{
synonyms <- psychsyn(data, .70)

antonyms <- psychsyn(data, anto=T) 
}
Last updated 22-june-2016
}
\author{
Richard Yentes \email(rdyentes@ncsu.edu)
}

