% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/psySyn.R
\name{psySyn}
\alias{psySyn}
\title{Psychological Synonyms}
\usage{
psySyn(x, cutoff = 0.6, cor.method = "pearson")
}
\arguments{
\item{x}{An \code{R} dataframe or matrix object.}

\item{cutoff}{An integer scalar. The value at which correlations should be declared "synonyms".
Default is set to \emph{r} <= +.60.}

\item{cor.method}{Character string indicating which method is passed to \code{cor} to use when
calculating the correlation coefficients.}
}
\value{
A vector of type numeric.
}
\description{
\code{psySyn} returns a numeric vector, pertaining to the individual's consistency index based
upon the correlation between items of a large magnitude.
}
\details{
This function is a consistency index based upon correlating all items within a survey, finding
the items with the largest correlations, and then calculating the within item correlation
among those measures.
}
\seealso{
\code{\link{careless}}, \code{\link{longString}}, \code{\link{psyAnt}}, \code{\link{malDist}}
}

