% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/withinsd.R
\name{withinsd}
\alias{withinsd}
\title{Within Person Standard Deviation}
\usage{
withinsd(x, na.rm = TRUE)
}
\arguments{
\item{x}{An \code{R} dataframe or matrix object.}

\item{na.rm}{A boolean string indicating whether cases with missing values should return a standard deviation based upon
the present values, or return \code{NA}.}
}
\value{
A numeric vector with the within person standard deviations.
}
\description{
\code{withinsd} returns the within person standard deviation for each row of a dataframe.
}
\seealso{
\code{\link{careless}}, \code{\link{psySyn}}
}

