% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/careless.R
\name{careless}
\alias{careless}
\title{Careless Summary}
\usage{
careless(x, append = TRUE, na.rm = FALSE, antonyms = FALSE,
  synonyms = FALSE, cutoff = 0.6)
}
\arguments{
\item{x}{An \code{R} dataframe or matrix object.}

\item{append}{A boolean scalar. If \code{append = TRUE} the original dataframe is returned with the
careless responding metrics bound to it. If \code{append = FALSE}, only the careless responding
metrics are returned.}

\item{na.rm}{A logical scalar. This is passed to \code{longString()} and \code{withinsd()}.}

\item{cutoff}{A numeric scalar. If \code{psySyn} or \code{psyAnt} are \code{TRUE},
\code{cutoff} is the absolute value at which correlations should be declared
"antonyms" and "synonyms". Default is set to \emph{r} +/- .60.}

\item{psySyn}{A boolean scalar. Should Psychological Synonyms be evaluated?}

\item{psyAnt}{A boolean scaler. Should Psychological Antonyms be evaluated?}
}
\value{
A dataframe object.
}
\description{
\code{careless} is a convenience function that calls the various methods within the package
and, by default, appends them onto the original dataframe. This function is designed to work with
dataframe or matrix objects.
}
\examples{
\dontrun{
set.seed(77)
dat <- matrix(sample(1:3, 200, replace = TRUE),20,10)
stats <- careless(dat, append = FALSE)
dat <- careless(dat, append = TRUE)
}
}
\seealso{
\code{\link{longString}},\code{\link{malDist}}
}

