% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/careless-package.R, R/careless.R
\docType{package}
\name{careless}
\alias{careless}
\alias{careless-package}
\title{careless: Metrics for identifying careless responding.}
\usage{
careless(x, append = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} dataframe or matrix object.}

\item{append}{A boolean scalar. If \code{append = TRUE} the original dataframe is returned with the
careless responding metrics bound to it. If \code{append = FALSE}, only the careless responding
metrics are returned.}

\item{na.rm}{A logical scalar. This is passed to \code{longString()}.}
}
\value{
A dataframe object.
}
\description{
The careless package provides a variety of metrics for identifying careless responding.
This these are drawn from the statistics used by Craig and Meade (2012).

\code{careless} is a summary function that calls various functions within the \code{careless} package.
}
\details{
This function is designed to work with a dataframe or matrix object.
}
\section{careless functions}{
 So far, this package implements the following functions:
`longString()`, which returns either the count or the value of the longest set of
repititions in a dataframe.
}
\examples{
\dontrun{
set.seed(77)
dat <- matrix(sample(1:3, 200, replace = TRUE),20,10)
stats <- careless(dat, append = FALSE)
dat <- careless(dat, append = TRUE)
}
}
\seealso{
\code{\link{longString}},\code{\link{malDist}}
}

