% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenodd.R
\name{evenodd}
\alias{evenodd}
\title{Calculates the even-odd consistency score}
\usage{
evenodd(d, factors)
}
\arguments{
\item{d}{is a matrix of data (e.g. survey repsonses)}

\item{factors}{is a vector of integers specifying the length of each
factor in the dataset}
}
\description{
Takes a matrix of item responses and a vector of integers representing the 
length each factor. The even-odd consistency score is then computed as the
within-person correlation between the even and odd subscales over all the 
factors
}
\examples{
\dontrun{
eo <- evenodd(d, c(rep(10,10)))
}
}
\author{
Richard Yentes \email(rdyentes@ncsu.edu)
}

